Move idUI to its own component
